chore: investigation into dialog usage
Implementing several approaches for the use of the <dialog> html
element, the goal was to get familiar with the element and it's api, as
well as explore possible options for no-js options for using it.

Overall the element is reasonable, and give a nicer, semantic tag for
a modal instead of a `<div aria_role="dialog"></div>`.

Positioning is tedious, based on margin sizes.
Animation is fine for displaying the modal, however it falls short on
closing the modal, as the browser is only concerned with the display
property and nothing else.

Without Javascript, you can achieve some modal-like functionality using
:target, :not, :has selectors, but I would not rely heavily on these for
complex operations. You could also use checkbox inputs and labels to
trigger the display for more stable control and without mangling the
browser history.

If building with progressive enhancement, the places I can see viable
use for <dialog> as a modal without the use of javascript are when
html is returned from the server and the open attribute is already
applied to the element <dialog open>. This then permits the dialog to be
closed with both <esc> keypress and using a form with method="dialog".